draft/evernotepy/requirements.txt
evernote
requests
requests_oauthlib
python-dotenv